# swagger.yaml
openapi: 3.0.3
info:
  title: Local Server Gateway
  description: Gateway для очереди задач между локальными устройствами
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Local

paths:
  /health:
    get:
      summary: Healthcheck
      tags: [System]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /tasks:
    get:
      summary: Список всех задач
      tags: [Tasks]
      responses:
        '200':
          description: Массив задач
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

    post:
      summary: Создать задачу
      tags: [Tasks]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  description: Тип задачи (определяет какой обработчик её возьмёт)
                  example: photo
                content:
                  description: Произвольные данные задачи (JSON-строка или plain string)
                  example: '{"album":"vacation","source":"iphone"}'
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Прикреплённые файлы
      responses:
        '201':
          description: Задача создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /tasks/{id}:
    get:
      summary: Получить задачу по ID
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Задача
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /files/{filename}:
    get:
      summary: Скачать загруженный файл
      tags: [Files]
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
          example: 550e8400-e29b-41d4-a716-446655440000.jpg
      responses:
        '200':
          description: Файл
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        type:
          type: string
          example: photo
        content:
          nullable: true
          description: Произвольные данные (зависит от типа задачи)
          example: { "album": "vacation" }
        files:
          type: array
          items:
            $ref: '#/components/schemas/TaskFile'
        provider:
          type: string
          nullable: true
          description: Socket ID обработчика, забравшего задачу
          example: abc123xyz
        status:
          type: string
          enum: [waiting, in_progress, failed]
          example: waiting
        date:
          type: string
          format: date-time
          description: Время последнего изменения
          example: '2024-01-01T12:00:00.000Z'
        error:
          type: string
          nullable: true
          description: Причина ошибки (только при status=failed)

    TaskFile:
      type: object
      properties:
        filename:
          type: string
          description: Имя файла на сервере (UUID)
          example: 550e8400-e29b-41d4-a716-446655440000.jpg
        originalname:
          type: string
          description: Оригинальное имя файла
          example: photo.jpg
        mimetype:
          type: string
          example: image/jpeg
        size:
          type: integer
          description: Размер в байтах
          example: 204800

    Error:
      type: object
      properties:
        error:
          type: string

  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: type is required
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Task not found
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error

tags:
  - name: Tasks
    description: Управление задачами
  - name: Files
    description: Загруженные файлы
  - name: System
    description: Системные ручки